nfs: implement nonce for mounting
You must specify the UUID as the path for mounting
and you only get one try.

The first is to ensure only people we want to mount mount;
the second to ensure nobody tries to guess.

The nonce will be held in the environment variable CPU_FSTAB,
and it's all over ssh, so there is a reasonable expectation
of privacy.

Signed-off-by: Ron Minnich <rminnich@gmail.com>