Lstat. But Stat() is worng for cpiofs, does not follow links
Later, we will see that Stat should not follow links,
it confuses kernel clients if we do.

Note that a billy example (mem) makes the mistake of following
a link, it follows it once and only once. That's wrong from
both sides: either do zero follows or none at all.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>