The time has come: default sidecore to nfs, leave 9p off
There are too many IoT that come with nfs working,
and built in; and 9p, either a module, or not even available
under any circumstances.

Even when 9p is available, its performance is bad enough that
it makes the overall experience very unpleasant. 30s to tab-complete
/usr/bin? No.

This breaks my heart, but it is time to accept reality. It's not
that 9p can not perform well. My first version, in 1998, could
hold its own with nfs. But it seems that such performance has
not been a priority. Time to accept reality.

Signed-off-by: Ron Minnich <rminnich@gmail.com>